app.py
import os
import numpy as np
from flask import Flask, render_template, request, jsonify, flash
from werkzeug.utils import secure_filename
from PIL import Image
import base64
import io
import random

app = Flask(__name__)
app.secret_key = 'your-secret-key-here'

# Configuration
UPLOAD_FOLDER = 'uploads'
ALLOWED_EXTENSIONS = {'png', 'jpg', 'jpeg', 'gif', 'bmp'}
MAX_FILE_SIZE = 16 * 1024 * 1024  # 16MB max file size

# CIFAR-10 class names
CIFAR10_CLASSES = [
    'airplane', 'automobile', 'bird', 'cat', 'deer',
    'dog', 'frog', 'horse', 'ship', 'truck'
]

# Ensure upload folder exists
os.makedirs(UPLOAD_FOLDER, exist_ok=True)

# Initialize model variable
model = None

# Try to load TensorFlow model, fallback to mock if not available
try:
	import tensorflow as tf
	from tensorflow.keras.models import load_model
	
	# Try to load the pre-trained model (prefer output/My-App-Model.h5)
	MODEL_CANDIDATES = [
		'output/My-App-Model.h5',
		'My-App-Model.h5',
		'cifar10_advanced_model.h5'
	]
	loaded = False
	for candidate in MODEL_CANDIDATES:
		if os.path.exists(candidate):
			model = load_model(candidate)
			print(f"[INFO] TensorFlow model loaded successfully from {candidate}.")
			loaded = True
			break
	if not loaded:
		print("[WARN] No model file found, using mock classification.")
		
except ImportError:
	print("[WARN] TensorFlow not available, using mock classification.")
except Exception as e:
	print(f"[WARN] Error loading model: {e}")
	print("[WARN] Using mock classification instead.")

def allowed_file(filename):
	return '.' in filename and \
		   filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS

def preprocess_image(image_path):
	"""Preprocess image for CIFAR-10 model input"""
	try:
		# Load and resize image to 32x32 (CIFAR-10 input size)
		img = Image.open(image_path)
		img = img.convert('RGB')
		img = img.resize((32, 32))
		
		# Convert to numpy array and normalize
		img_array = np.array(img)
		img_array = img_array.astype('float32') / 255.0
		
		# Add batch dimension
		img_array = np.expand_dims(img_array, axis=0)
		
		return img_array
	except Exception as e:
		print(f"Error preprocessing image: {e}")
		return None

def mock_classification():
	"""Mock classification when TensorFlow is not available"""
	# Generate random but realistic-looking predictions
	predictions = []
	used_classes = set()
	
	for i in range(3):
		# Get a random class that hasn't been used yet
		available_classes = [c for c in CIFAR10_CLASSES if c not in used_classes]
		if not available_classes:
			available_classes = CIFAR10_CLASSES
			
		class_name = random.choice(available_classes)
		used_classes.add(class_name)
		
		# Generate realistic confidence scores (decreasing for each rank)
		if i == 0:  # Top prediction
			confidence = random.uniform(0.6, 0.95)
		elif i == 1:  # Second prediction
			confidence = random.uniform(0.2, 0.6)
		else:  # Third prediction
			confidence = random.uniform(0.05, 0.3)
		
		predictions.append({
			'class': class_name,
			'confidence': confidence,
			'percentage': round(confidence * 100, 2)
		})
	
	# Sort by confidence (highest first)
	predictions.sort(key=lambda x: x['confidence'], reverse=True)
	return predictions

def predict_image(image_path):
	"""Make prediction on the uploaded image"""
	if model is None:
		# Use mock classification
		print("Using mock classification (TensorFlow model not available)")
		predictions = mock_classification()
		return predictions, None
	
	try:
		# Preprocess the image
		processed_img = preprocess_image(image_path)
		if processed_img is None:
			return None, "Error preprocessing image"
		
		# Make prediction with real model
		predictions = model.predict(processed_img)
		
		# Get top 3 predictions
		top_indices = np.argsort(predictions[0])[-3:][::-1]
		top_predictions = []
		
		for idx in top_indices:
			class_name = CIFAR10_CLASSES[idx]
			confidence = float(predictions[0][idx])
			top_predictions.append({
				'class': class_name,
				'confidence': confidence,
				'percentage': round(confidence * 100, 2)
			})
		
		return top_predictions, None
		
	except Exception as e:
		print(f"Real model prediction failed: {e}")
		print("Falling back to mock classification")
		predictions = mock_classification()
		return predictions, None

@app.route('/')
def index():
	return render_template('index.html')

@app.route('/upload', methods=['POST'])
def upload_file():
	if 'file' not in request.files:
		return jsonify({'error': 'No file part'}), 400
	
	file = request.files['file']
	
	if file.filename == '':
		return jsonify({'error': 'No selected file'}), 400
	
	if file and allowed_file(file.filename):
		try:
			# Check file size
			file.seek(0, os.SEEK_END)
			file_size = file.tell()
			file.seek(0)
			
			if file_size > MAX_FILE_SIZE:
				return jsonify({'error': 'File too large. Maximum size is 16MB.'}), 400
			
			# Save file
			filename = secure_filename(file.filename)
			filepath = os.path.join(UPLOAD_FOLDER, filename)
			file.save(filepath)
			
			# Make prediction
			predictions, error = predict_image(filepath)
			
			if error:
				return jsonify({'error': error}), 500
			
			# Convert image to base64 for display
			with open(filepath, 'rb') as img_file:
				img_data = base64.b64encode(img_file.read()).decode('utf-8')
			
			# Clean up uploaded file
			os.remove(filepath)
			
			return jsonify({
				'success': True,
				'predictions': predictions,
				'image_data': img_data,
				'filename': filename,
				'model_loaded': model is not None
			})
			
		except Exception as e:
			return jsonify({'error': f'Error processing file: {str(e)}'}), 500
	
	return jsonify({'error': 'Invalid file type. Allowed: PNG, JPG, JPEG, GIF, BMP'}), 400

@app.route('/health')
def health_check():
	return jsonify({
		'status': 'healthy', 
		'model_loaded': model is not None,
		'tensorflow_available': 'tensorflow' in globals(),
		'model_file_exists': any(os.path.exists(p) for p in ['output/My-App-Model.h5', 'My-App-Model.h5', 'cifar10_advanced_model.h5'])
	})

@app.route('/status')
def status():
	"""Detailed status endpoint for debugging"""
	candidate = (
		'output/My-App-Model.h5' if os.path.exists('output/My-App-Model.h5') else (
			'My-App-Model.h5' if os.path.exists('My-App-Model.h5') else 'cifar10_advanced_model.h5'
		)
	)
	status_info = {
		'flask_version': '2.3.3',
		'python_version': '3.13.6',
		'model_loaded': model is not None,
		'tensorflow_available': 'tensorflow' in globals(),
		'model_candidate': candidate if os.path.exists(candidate) else None,
		'model_file_exists': os.path.exists(candidate),
		'model_file_size': os.path.getsize(candidate) if os.path.exists(candidate) else 0,
		'upload_folder_exists': os.path.exists(UPLOAD_FOLDER),
		'working_mode': 'real_model' if model is not None else 'mock_classification'
	}
	return jsonify(status_info)

if __name__ == '__main__':
	print("[START] Starting CIFAR-10 Image Classifier...")
	candidate = (
		'output/My-App-Model.h5' if os.path.exists('output/My-App-Model.h5') else (
			'My-App-Model.h5' if os.path.exists('My-App-Model.h5') else 'cifar10_advanced_model.h5'
		)
	)
	print(f"[INFO] Preferred model file: {candidate}")
	print(f"[INFO] TensorFlow model loaded: {model is not None}")
	print(f"[INFO] Working mode: {'Real AI Model' if model is not None else 'Mock Classification'}")
	print("[INFO] Starting Flask server at http://localhost:5000")
	
	app.run(debug=True, host='0.0.0.0', port=5000)
