requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow>=10.0.0
numpy>=1.24.0
